Completed Activity 2
